Improve Gradle plugin written in Java sample
- Use com.example namespace instead of org.gradle
- Run the plugin tests
- Remove some comments
- Rename the plugin to GreetingPlugin
