+++ subprojects/docs/src/samples/gradle-plugin/tests/check.sample.conf
@@ -1,0 +1,4 @@
+commands: [{
+    executable: gradle
+    args: check
+}]
