--- subprojects/docs/src/samples/templates/gradle-plugin-in-java/src/main/java/org/gradle/sample/plugin/GradlePluginInJavaPlugin.java
@@ -1,19 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package org.gradle.sample.plugin;
-
-import org.gradle.api.Project;
-import org.gradle.api.Plugin;
-
-/**
- * A simple 'hello world' plugin.
- */
-public class GradlePluginInJavaPlugin implements Plugin<Project> {
-    public void apply(Project project) {
-        // Register a task
-        project.getTasks().register("greeting", task -> {
-            task.doLast(s -> System.out.println("Hello from plugin 'org.gradle.sample.plugin.greeting'"));
-        });
-    }
-}
